Acl accesslist (#80)
* first cut of accesslist processing.

* adding compilation steps.

* adding compilation steps.

* adding placeholders for accesslist whitebox tests.

* updating is2.

* accesslist fixes and tests.

* removing dead code.

* fixes for waflz_server running in limit mode.

* cleanup from -Weffc++ and -Wall.

* cleanup from -Weffc++ and -Wall.

* cleanup from -Weffc++ and -Wall.

* cleanup from -Weffc++ and -Wall.

* cleanup from -Weffc++ and -Wall.

* cleanup from -Weffc++ and -Wall.

* cleanup from -Weffc++ and -Wall.

* cleanup from -Weffc++ and -Wall.

* cleanup from -Weffc++ and -Wall.

* cleanup from -Weffc++ and -Wall.

* cleanup from -Weffc++ and -Wall.

* cleanup from -Weffc++ and -Wall.

* more acl_accesslists testing.

* cleanup from -Weffc++ and -Wall.

* accesslists tess.

* accesslists tess.

* adding placeholders for accesslists tests.

+++ include/waflz/rqst_ctx.h
@@ -148,7 +148,7 @@ public:
         const uint32_t m_body_len_max;
         char *m_body_data;
         uint32_t m_body_len;
-        uint32_t m_content_length;
+        uint64_t m_content_length;
         bool m_parse_xml;
         bool m_parse_json;
         std::string m_cookie_mutated;
